requirements.txt
streamlit==1.39.0
PyMuPDF==1.24.9
pdfminer.six==20231228
pdfplumber==0.11.4
pypdfium2==4.30.0
PyPDF2==3.0.1
pytesseract==0.3.13
numpy==1.26.4
opencv-python-headless==4.10.0.84
easyocr==1.7.2
requests==2.32.3
Pillow==10.2.0
google-cloud-vision==3.7.4
langchain==0.3.1
llama-index-core==0.11.17
llama-index-readers-file==0.2.2
llama-parse==0.5.7
python-dotenv==1.0.1
nest-asyncio==1.6.0
pdf2image==1.17.0
python-magic==0.4.27